final-pjt-back/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==3.2.12
django-allauth==0.49.0
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
idna==3.3
joblib==1.1.0
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
urllib3==1.26.8
